fixing install modules.

+++ setup/utils/install_play.sh
@@ -11,8 +11,10 @@ fi
 if [ -f /usr/bin/play ]; then
     echo "play already exists - nothing to do"
 else
-    wget 'http://download.playframework.org/releases/play-2.0.4.zip'  -O /tmp/play.zip
-    unzip -o /tmp/play.zip -d /usr/lib
+    if [ ! -f /tmp/play.zip ]; then
+        wget 'http://download.playframework.org/releases/play-2.0.4.zip'  -O /tmp/play.zip
+    fi
+    unzip -q -o /tmp/play.zip -d /usr/lib
     ln -Tfs ${PLAY_HOME} /usr/lib/play
     rm -f /tmp/play.zip
 fi
